backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pymupdf>=1.24.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
